refactor(runtypes): name the batch extractor package requestbatch
A bare `batches` said nothing about what it batches; this is the first marker
lane that serves mion's RPC layer rather than RunTypes itself.
