+++ ts-go-runtypes/internal/compiler/requestbatch/batches_test.go
@@ -1,4 +1,4 @@
-package batches
+package requestbatch
 
 import (
 	"context"
