+++ ts-go-runtypes/internal/compiler/requestbatch/discover.go
@@ -1,4 +1,4 @@
-package batches
+package requestbatch
 
 import (
 	"github.com/microsoft/typescript-go/shim/ast"
