+++ ts-go-runtypes/internal/compiler/requestbatch/report.go
@@ -1,4 +1,4 @@
-package batches
+package requestbatch
 
 import (
 	"sort"
